refactor: use client form for set password
