+++ app/set-password/ClientForm.tsx
@@ -1,96 +1,110 @@
-// app/set-password/page.tsx
 "use client";
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-import { signIn, useSession } from "next-auth/react";
-import PasswordInput from "@/components/PasswordInput";
 
-export default function SetPasswordPage() {
-  const { data: session } = useSession();
+import { useState } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { FaEye, FaEyeSlash } from "react-icons/fa";
+
+export default function ClientForm() {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const token = searchParams.get("token");
+
   const [password, setPassword] = useState("");
   const [confirmPassword, setConfirmPassword] = useState("");
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState("");
+  const [showPassword, setShowPassword] = useState(false);
+  const [showConfirm, setShowConfirm] = useState(false);
+  const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
-  const router = useRouter();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    setError("");
-    setSuccess("");
-    setLoading(true);
+    setError(null);
+
+    if (!token) {
+      setError("Missing token.");
+      return;
+    }
 
     if (password.length < 6) {
       setError("Password must be at least 6 characters.");
-      setLoading(false);
       return;
     }
 
     if (password !== confirmPassword) {
       setError("Passwords do not match.");
-      setLoading(false);
       return;
     }
 
-    const res = await fetch("/api/set-password", {
-      method: "POST",
-      body: JSON.stringify({ password }),
-      headers: { "Content-Type": "application/json" },
-    });
-
-    const data = await res.json();
-
-    if (!res.ok) {
-      setError(data.error || "Something went wrong.");
-    } else {
-      setSuccess("Password updated. Redirecting...");
-      await signIn("credentials", {
-        redirect: false,
-        email: session?.user.email!,
-        password,
+    setLoading(true);
+    try {
+      const res = await fetch("/api/set-password", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ token, password }),
       });
-      router.push("/profile");
+      if (!res.ok) {
+        const data = await res.json().catch(() => ({}));
+        setError(data.error || "Failed to set password.");
+      } else {
+        router.push("/profile");
+      }
+    } catch {
+      setError("Failed to set password.");
+    } finally {
+      setLoading(false);
     }
-
-    setLoading(false);
   };
 
   return (
-    <div className="max-w-md mx-auto mt-10 p-6 bg-white shadow rounded">
-      <h1 className="text-xl font-bold mb-4">Set a New Password</h1>
-      <p className="text-sm text-gray-600 mb-4">
-        This will allow you to log in using your email and password in the future.
-      </p>
+    <div className="max-w-md mx-auto py-12">
+      <h1 className="text-2xl font-bold mb-4">Set Password</h1>
       <form onSubmit={handleSubmit} className="space-y-4">
-        <div>
-          <label className="block mb-1 font-medium">New Password</label>
-          <PasswordInput
+        <div className="relative">
+          <input
+            type={showPassword ? "text" : "password"}
             value={password}
             onChange={(e) => setPassword(e.target.value)}
-            className="w-full border px-4 py-2 rounded"
+            placeholder="New Password"
             required
             minLength={6}
+            className="w-full border px-4 py-2 rounded pr-10"
           />
+          <button
+            type="button"
+            onClick={() => setShowPassword((s) => !s)}
+            className="absolute right-2 top-1/2 -translate-y-1/2 text-gray-600"
+            aria-label={showPassword ? "Hide password" : "Show password"}
+          >
+            {showPassword ? <FaEyeSlash /> : <FaEye />}
+          </button>
         </div>
 
-        <div>
-          <label className="block mb-1 font-medium">Confirm Password</label>
-          <PasswordInput
+        <div className="relative">
+          <input
+            type={showConfirm ? "text" : "password"}
             value={confirmPassword}
             onChange={(e) => setConfirmPassword(e.target.value)}
-            className="w-full border px-4 py-2 rounded"
+            placeholder="Confirm Password"
             required
             minLength={6}
+            className="w-full border px-4 py-2 rounded pr-10"
           />
+          <button
+            type="button"
+            onClick={() => setShowConfirm((s) => !s)}
+            className="absolute right-2 top-1/2 -translate-y-1/2 text-gray-600"
+            aria-label={showConfirm ? "Hide password" : "Show password"}
+          >
+            {showConfirm ? <FaEyeSlash /> : <FaEye />}
+          </button>
         </div>
 
         {error && <p className="text-red-600 text-sm">{error}</p>}
-        {success && <p className="text-green-600 text-sm">{success}</p>}
 
         <button
           type="submit"
           disabled={loading}
-          className="bg-blue-600 text-white px-4 py-2 rounded disabled:opacity-50"
+          className="bg-blue-600 text-white px-4 py-2 rounded w-full disabled:opacity-50"
         >
           {loading ? "Saving..." : "Set Password"}
         </button>
